Move util methods to appropriate classes.
